fix(discussions): filter the composer's space picker the way the dialog does
useNewDiscussion filtered spaces on archived_at and the route's community only. No
guest check, no read-only check. So it offered a guest a space to post in that the
backend would refuse: can_create_content lets a guest create a GP Comment or a GP
Poll and nothing else, in every space including the ones they were granted.

A guest cannot create a discussion draft, so there is no way in here from the drafts
list. Typing the composer URL is the only reachable path, which makes this low
severity rather than none. It is still worth closing: canPostInSpace is now the
shared predicate for this question, and a second filter answering it differently is
how the hole survives the next change. The composer asks canPostInSpace too.

A user left with nothing to pick used to get an empty picker and no explanation. The
composer now shows the same empty state the space dialog does, moved into a shared
NoSpaceToPostIn component so the wording cannot drift between the two.
